[Docs](fix) Repair A5 libdevice API example tests from #1782

+++ docs/zh/python-api/_examples/triton.language.extra.cann.libdevice.ull2float_rz.py
@@ -25,7 +25,7 @@ def triton_kernel(input0, output, n_elements, XBLOCK: tl.constexpr, XBLOCK_SUB: 
         x0 = offset + (loop * XBLOCK_SUB) + base
         mask = x0 < n_elements
         tmp0 = tl.load(input0 + (x0), mask=mask)
-        tmp1 = libdevice.ull2float_rz(tmp0)
+        tmp1 = libdevice.ull2float_rz(tmp0.to(tl.uint64))
         tl.store(output + (x0), tmp1, mask=mask)
